add explicit casts to all max_n_rels with value null to increase readability and to ease function disambiguation

+++ sql/postgresql/admins-init.sql
@@ -43,7 +43,7 @@ begin
         ''profiled_group'',
         null,
         0,
-        null,
+        null::integer,
         ''user'',
         null,
         0,
