+++ sql/postgresql/advisers-init.sql
@@ -42,7 +42,7 @@ begin
         ''profiled_group'',
         null,
         0,
-        null,
+        null::integer,
         ''user'',
         null,
         0,
