+++ sql/postgresql/users-init.sql
@@ -40,7 +40,7 @@ begin
         ''profiled_group'',
         null,
         0,
-        null,
+        null::integer,
         ''user'',
         null,
         0,
